Added comment, created Util class
